pyproject.toml
[project]
name = "sql-tools"
version = "0.0.3"
description = "A collection of utility tools for working with various dialects of SQL databases."
license = { file = "LICENSE" }
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyodbc",
    "psycopg2-binary",
    "python-dotenv",
    "rich",
    "toml",
    "sqlalchemy",
    "pandas",
    "pyarrow",
]

[project.optional-dependencies]
dev = ["ruff", "mypy", "types-toml", "pandas-stubs", "types-psycopg2"]

[project.scripts]
object_compare = "object_compare.object_compare:main"
usp_tester = "usp_tester.usp_tester:main"
view_tester = "view_tester.view_tester:main"
sql_to_parquet = "sql_to_parquet.sql_to_parquet:main"
schema_size = "schema_size.schema_size:main"
data_compare = "data_compare.data_compare:main"
data_cleanup = "data_cleanup.data_cleanup:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "./src" }

[tool.ruff]
target-version = "py313"
indent-width = 4
line-length = 100
exclude = ["*.json"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "W", "C90"]

[tool.ruff.format]
indent-style = "space"
line-ending = "auto"
quote-style = "double"

[tool.mypy]
mypy_path = "src"
disallow_untyped_defs = true
python_version = "3.13"
warn_return_any = true
strict = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
explicit_package_bases = true
namespace_packages = true
exclude = ["temp/.*"]

[[tool.mypy.overrides]]
module = ["pyodbc"]
ignore_missing_imports = true

[tool.pyright]
venvPath = "."
venv = ".venv"
